% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.box.R
\name{data_arange}
\alias{data_arange}
\title{Format TPM data}
\usage{
data_arange(file_path, gene_set, sample, replication)
}
\arguments{
\item{file_path}{TPMファイルのパス}

\item{gene_set}{発現を比較したい遺伝子。ベクトル形式で入力する。}

\item{sample}{系統名。ベクトル形式で入力する}

\item{replication}{反復数}
}
\description{
Format TPM data
}
