% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.box.R
\name{expression_box}
\alias{expression_box}
\title{Create Boxplot to compare gene expression between different populations}
\usage{
expression_box(file_path = "TPM.txt", gene_set, sample, replication = 3, pal)
}
\arguments{
\item{file_path}{TPMファイルのパス}

\item{gene_set}{発現を比較したい遺伝子。ベクトル形式で入力する。}

\item{sample}{系統名。ベクトル形式で入力する}

\item{replication}{反復数}

\item{pal}{ボックスの色。系統数の色をベクトル形式で入力する。}
}
\description{
Create Boxplot to compare gene expression between different populations
}
