% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.box.R
\name{generate_both_table}
\alias{generate_both_table}
\title{Create both of count table and TPM table}
\usage{
generate_both_table(directory_path = "result")
}
\arguments{
\item{directory_path}{RSEMで出力される.resultファイルがあるディレクトリのパス}
}
\description{
Create both of count table and TPM table
}
