% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.box.R
\name{generate_exp_table}
\alias{generate_exp_table}
\title{Create count table}
\usage{
generate_exp_table(directory_path = "result")
}
\arguments{
\item{directory_path}{RSEMで出力される.resultファイルがあるディレクトリのパス}
}
\description{
Create count table
}
