% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.box.R
\name{generate_TPM_table}
\alias{generate_TPM_table}
\title{Create TPM table}
\usage{
generate_TPM_table(directory_path)
}
\arguments{
\item{directory_path}{RSEMで出力される.resultファイルがあるディレクトリのパス}
}
\description{
Create TPM table
}
